did some restructuring, moved knexconfig to zoos-router, server running, starting endpoints
