+++ zoos/zoos-router.js
@@ -1,0 +1,15 @@
+const router = require("express").Router();
+const knex = require("knex");
+
+const knexConfig = {
+  client: "sqlite3",
+  useNullAsDefault: true,
+  connection: {
+    filename: "./data/lambda.sqlite3"
+  },
+  debug: true
+};
+
+const db = knex(knexConfig);
+
+module.exports = router;
